src/main/java/com/example/handle/localdate/birthday/BirthdayController.java
package com.example.handle.localdate.birthday;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

import static org.springframework.http.HttpStatus.CREATED;
import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
import static org.springframework.web.bind.annotation.RequestMethod.GET;
import static org.springframework.web.bind.annotation.RequestMethod.POST;

@RestController
public class BirthdayController {

    private BirthdayService birthdayService;

    @Autowired
    public BirthdayController(BirthdayService birthdayService) {
        this.birthdayService = birthdayService;
    }

    @RequestMapping(value = "/birthday/create", method = POST, consumes = APPLICATION_JSON_VALUE)
    @ResponseStatus(CREATED)
    public void create(@RequestBody Birthday birthday) {}

    @RequestMapping(value = "/birthdays", method = GET, produces = APPLICATION_JSON_VALUE)
    public List<Birthday> findAll() {
        return birthdayService.findAll();
    }

}
